// src/app/Components/login-design/swt-alert.enum.ts
export enum AlertTitle {
  CreateTitle = 'Created!',
  UpdateTitle = 'Updated!',
  Deletetitle = 'Deleted!',
  WarningTitle = 'Are You Sure?',
  SaveMessage = "SaveMessage"
}
export enum AlertMessage {
    Message= '',
 }
export enum AlertIcon {
  WarningIcon = 'warning',
  SuccessIcon = 'success',
  UpdateIcon = 'success',
  ErrorIcon = 'error',
  DeleteIcon = 'success',
  DeleteWarningIcon = 'warning'
}
 export enum AlertMessage {
   CreateMessage= 'Created Successfully!',
   UpdateMessage= 'Updated Successfully!',
   DeleteMessage= 'Deleted Successfully!',
   SaveMessage='Saved Successfully!',
      DeleteWarningIcon='warning'}